pyproject.toml
[tool.poetry]
name = "tenant"
version = "0.1.0"
description = ""
authors = ["mgur <mgur@pop-os.localdomain>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.39"
alembic = "^1.8.0"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
loguru = "^0.6.0"
typeguard = "^2.13.3"
Faker = "^13.15.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.6.0"
pytest-coverage = "^0.0"
isort = "^5.10.1"
autoflake = "^1.4"
sqlacodegen = "3.0.0rc1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
